requirements.txt
accelerate==1.5.2
bitsandbytes==0.46.1
datasets==3.6.0
huggingface-hub==0.31.1
numpy==1.26.4
pandas==2.2.3
peft==0.14.0
PyYAML==6.0.2
safetensors==0.5.3
scikit-learn==1.2.2
torch @ https://download.pytorch.org/whl/cu124/torch-2.6.0%2Bcu124-cp311-cp311-linux_x86_64.whl
torchvision @ https://download.pytorch.org/whl/cu124/torchvision-0.21.0%2Bcu124-cp311-cp311-linux_x86_64.whl
tqdm==4.67.1
transformers==4.51.3



